Move `screens` package to appropriate locations
