+++ app/src/main/java/com/it235/nureserved/MainActivity.kt
@@ -22,17 +22,17 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.google.firebase.auth.FirebaseAuth
-import com.it235.nureserved.ui.screens.RoomReservationForm
-import com.it235.nureserved.ui.screens.RoomUsageRules
-import com.it235.nureserved.ui.screens.SplashScreen
-import com.it235.nureserved.ui.screens.TermsAndConditionsScreen
-import com.it235.nureserved.ui.screens.authscreenui.LoginScreen
-import com.it235.nureserved.ui.screens.authscreenui.signup.NameSignUpScreen
-import com.it235.nureserved.ui.screens.authscreenui.signup.ProgramStudentNumberSignUpScreen
-import com.it235.nureserved.ui.screens.authscreenui.signup.SignUpScreen
-import com.it235.nureserved.ui.screens.homesreenui.HomeScreen
-import com.it235.nureserved.ui.screens.homesreenui.RoomDetails
-import com.it235.nureserved.ui.screens.onboardingscreenui.GetStartedScreen
+import com.it235.nureserved.screens.user.RoomReservationForm
+import com.it235.nureserved.screens.user.RoomUsageRules
+import com.it235.nureserved.screens.core.SplashScreen
+import com.it235.nureserved.screens.user.TermsAndConditionsScreen
+import com.it235.nureserved.screens.core.authscreenui.LoginScreen
+import com.it235.nureserved.screens.core.authscreenui.signup.NameSignUpScreen
+import com.it235.nureserved.screens.core.authscreenui.signup.ProgramStudentNumberSignUpScreen
+import com.it235.nureserved.screens.core.authscreenui.signup.SignUpScreen
+import com.it235.nureserved.screens.user.homesreenui.HomeScreen
+import com.it235.nureserved.screens.user.homesreenui.RoomDetails
+import com.it235.nureserved.screens.core.onboardingscreenui.GetStartedScreen
 import com.it235.nureserved.ui.theme.NUreservedTheme
 import kotlinx.coroutines.delay
