+++ app/src/main/java/com/it235/nureserved/screens/user/RoomUsageRulesScreen.kt
@@ -1,4 +1,4 @@
-package com.it235.nureserved.ui.screens
+package com.it235.nureserved.screens.user
 
 
 import androidx.compose.foundation.layout.Arrangement
